moved Reaction to own file
